pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dmaf"
version = "0.1.0"
description = "Don't Miss A Face — Automated WhatsApp media backup with intelligent face recognition filtering"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Yonatan"}
]
keywords = [
    "face-recognition",
    "whatsapp", 
    "google-photos",
    "photo-backup",
    "automation",
    "insightface",
    "dlib",
    "family-photos",
    "media-management",
    "privacy",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Home Automation",
]

dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "PyYAML>=6.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "requests>=2.28.0",
    "watchdog>=3.0.0",
    "google-auth>=2.22.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
]

[project.optional-dependencies]
face-recognition = [
    "face-recognition>=1.3.0",
    "dlib>=19.24.0",
]
insightface = [
    "insightface>=0.7.3",
    "onnxruntime>=1.15.0",
]
all = [
    "dmaf[face-recognition,insightface]",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
dmaf = "dmaf.__main__:main"

[project.urls]
Homepage = "https://github.com/yonatan/dmaf"
Documentation = "https://github.com/yonatan/dmaf#readme"
Repository = "https://github.com/yonatan/dmaf.git"
Issues = "https://github.com/yonatan/dmaf/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/dmaf"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=dmaf --cov-report=term-missing"
